% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIEprepExtra-package.R
\docType{package}
\name{DAISIEprepExtra-package}
\alias{DAISIEprepExtra}
\alias{DAISIEprepExtra-package}
\title{DAISIEprepExtra: Scripts and documents to accompany and reproduce Lambert et al. application paper describing the DAISIEprep R package}
\description{
This is a data and scripts package that stores the materials needed for reproduction of the examples, analyses and plots for the applications paper describing the DAISIEprep package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/joshwlambert/DAISIEprepExtra}
  \item Report bugs at \url{https://github.com/joshwlambert/DAISIEprepExtra/issues}
}

}
\author{
\strong{Maintainer}: Joshua W. Lambert \email{j.w.l.lambert@rug.nl} (\href{https://orcid.org/0000-0001-5218-3046}{ORCID})

Authors:
\itemize{
  \item Luis Valente \email{luis.valente@naturalis.nl} (\href{https://orcid.org/0000-0003-4247-8785}{ORCID})
  \item Lizzie Roeble \email{lizzie.roeble@naturalis.nl} (\href{https://orcid.org/0000-0003-3664-4222}{ORCID})
}

}
\keyword{internal}
