% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hawaii_asters}
\alias{hawaii_asters}
\title{Endemicity data on the Asteraceae species of Hawaii}
\format{
An object of class \code{data.frame} of length 126:
\describe{
  \item{tip_labels}{Species names at the tip of the phylogeny}
  \item{tip_endemicity_status}{Endemicity status of the species}
}
}
\usage{
hawaii_asters
}
\description{
A dataset containing the tip labels of the species in the tree and their
endemicity status on the archipelago of Hawaii.
}
\keyword{datasets}
