% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{madagascar_mammals}
\alias{madagascar_mammals}
\title{Endemicity data on the Mammals species of Madagascar}
\format{
An object of class \code{data.frame} of length 215:
\describe{
  \item{tip_labels}{Species names at the tip of the phylogeny}
  \item{tip_endemicity_status}{Endemicity status of the species}
}
}
\usage{
madagascar_mammals
}
\description{
A dataset containing the tip labels of the species in the tree and their
endemicity status on the island of Madagascar.
}
\keyword{datasets}
