% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world_border}
\alias{world_border}
\title{World map border in Robinson projection}
\format{
An object of class \code{SpatialPolygonDataFrame}
}
\usage{
world_border
}
\description{
An S4 SpatialPolygonDataFrame class containing data to plot the border of
the world map in a Robinson projection.
}
\keyword{datasets}
